1.Added delay in creation of bastion service to allow bastion agent on VM to come up first before creating the service.
2.Changed the bootstrap script to install Ruby without rbenv. The source command does not work with non-interactive shells
so installation of rbenv was failing. I think I have a workaround for this. I’ll test it and add it to the next update.

3.Changed sudo rails s -p 8080 -b 0.0.0.0 & to be a service instead. The terraform remote-exec shell works differently
than typical ssh. When the remote-exec dies so does all processes it spawns. I tried nohup and disown.
Changing the start of the service to a systemctl fixed the issue.